refactored help naming, added icon for fab for help guide
