+++ components/HelpGuide.module.css
@@ -1,0 +1,17 @@
+.fabStyle {
+    display: inline-flex;
+    align-items: center;
+    box-sizing: border-box;
+    cursor: pointer;
+    position: fixed;
+    bottom: 24px;
+    right: 24px;
+    z-index: 10;
+    color: white;
+    background-color: #3B5DD9;
+}
+
+.iconStyle {
+    width: 40px;
+    height: 40px;
+}
