+++ components/LightStatus.module.css
